+++ formula/src/test/java/com/instacart/formula/subjects/DynamicParentFormula.kt
@@ -4,6 +4,7 @@ import com.instacart.formula.Evaluation
 import com.instacart.formula.Formula
 import com.instacart.formula.FormulaContext
 import com.instacart.formula.Listener
+import com.instacart.formula.Snapshot
 
 class DynamicParentFormula(
     private val childFormula: KeyFormula = KeyFormula()
@@ -22,11 +23,7 @@ class DynamicParentFormula(
 
     override fun initialState(input: Unit): State = State()
 
-    override fun evaluate(
-        input: Unit,
-        state: State,
-        context: FormulaContext<State>
-    ): Evaluation<Output> {
+    override fun Snapshot<Unit, State>.evaluate(): Evaluation<Output> {
         return Evaluation(
             output = Output(
                 children = state.childKeys.map {