+++ samples/counter/src/main/java/com/instacart/formula/counter/CounterFormula.kt
@@ -3,16 +3,13 @@ package com.instacart.formula.counter
 import com.instacart.formula.Evaluation
 import com.instacart.formula.Formula
 import com.instacart.formula.FormulaContext
+import com.instacart.formula.Snapshot
 
 class CounterFormula : Formula<Unit, Int, CounterRenderModel>() {
 
     override fun initialState(input: Unit): Int = 0
 
-    override fun evaluate(
-        input: Unit,
-        state: Int,
-        context: FormulaContext<Int>
-    ): Evaluation<CounterRenderModel> {
+    override fun Snapshot<Unit, Int>.evaluate(): Evaluation<CounterRenderModel> {
         return Evaluation(
             output = CounterRenderModel(
                 count = "Count: $state",