+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - **Breaking**: Rename and move `FormulaContext.UpdateBuilder` into `StreamBuilder`. 
 - **Breaking**: Restructuring `Transition` type into `Transition<State, Event>` and `Transition.Result<State>`
 - **Breaking**: Make `Formula` abstract class instead of an interface.
+- **Breaking**: Introducing `Snapshot` type and changing `evaluate` signature.
 
 ## [0.7.0] - June 30, 2021
 - **Breaking**: Remove `events(observable) { }` utility function.